// payload-charity-starter/src/payload-types.ts
export interface Media {
  id: string;
  alt: string;
  updatedAt: string;
  createdAt: string;
  url?: string | null;
  thumbnailURL?: string | null;
  filename?: string | null;
  mimeType?: string | null;
  filesize?: number | null;
  width?: number | null;
  height?: number | null;
  focalX?: number | null;
  focalY?: number | null;
}
/**
 * This interface was referenced by `Config`'s JSON-Schema
 * via the `definition` "organizations".
 */
export interface Organization {
  id: string;
  name: string;
  logo: string | Media;
  updatedAt: string;
  createdAt: string;
}
// ...
export interface Campaign {
  id: string;
  organization: string | Organization;
  title: string;
  description: string;
  image: string | Media;
  updatedAt: string;
  createdAt: string;
}